+++ src/StateSmith/Output/UserConfig/IRenderConfigCSharp.cs
@@ -6,9 +6,16 @@ public interface IRenderConfigCSharp : IRenderConfig
     string Usings => "";
 
     /// <summary>
-    /// Use to add custom code to generated state machine class
+    /// Use to add custom code to generated state machine class, although <seealso cref="UsePartialClass"/> this
+    /// may suit your needs better.
     /// </summary>
     string ClassCode => "";
 
     bool UseNullable => true;
+
+    /// <summary>
+    /// Output state machine code will be a partial class. Very useful so you can define other part of 
+    /// partial class and easily add functionality to the state machine(fields, methods...).
+    /// </summary>
+    bool UsePartialClass => true;
 }